If the routing class is higher than 3 don't wait a full ogm cycle until a gateway is chosen. We change the gateway as soon as we find a better one, so no need to wait.
git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@1113 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ profile.h
@@ -40,12 +40,12 @@ struct prof_container {
 	clock_t start_time;
 	clock_t total_time;
 	char *name;
-	uint32_t calls;
+	uint64_t calls;
 
 };
 
 
-void prof_init( int32_t index, char *name );
-void prof_start( int32_t index );
-void prof_stop( int32_t index );
+void prof_init(int32_t index, char *name);
+void prof_start(int32_t index);
+void prof_stop(int32_t index);
 void prof_print();
